feat(release): slim the published manifest — string exports, drop devDeps/scripts
Rename prune-custom-export-conditions.js to clean-publish-manifest.js and
extend it beyond pruning custom conditions:

- collapse {types, import} export entries to plain string targets when
  types is the .d.ts sibling of import (TypeScript resolves it automatically)
- drop import conditions that duplicate default (the CSS entries)
- delete devDependencies and scripts from the published manifest

Published manifest shrinks 7,615 -> 3,987 bytes (mantle), roughly halving
the metadata each release permanently adds to the npm packument. attw
results on the packed tarball are identical before/after across node10,
node16-cjs, node16-esm, and bundler for all 68 subpaths. pnpm resolves the
pack lifecycle from the manifest read at startup, so deleting scripts in
prepack does not prevent postpack from running (verified empirically).
