+++ packages/mantle-server-syntax-highlighter/package.json
@@ -24,8 +24,8 @@
   },
   "scripts": {
     "build": "tsdown",
-    "prepack": "node ../../scripts/prune-custom-export-conditions.js prune",
-    "postpack": "node ../../scripts/prune-custom-export-conditions.js restore",
+    "prepack": "node ../../scripts/clean-publish-manifest.js clean",
+    "postpack": "node ../../scripts/clean-publish-manifest.js restore",
     "test": "TZ=UTC vitest run",
     "test:watch": "TZ=UTC vitest watch",
     "typecheck": "tsgo"
